requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
openai>=1.0.0
python-docx>=1.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
nbconvert>=7.0.0
nbformat>=5.9.0
aiofiles>=23.0.0
rarfile>=4.0
py7zr>=0.20.0
websockets>=12.0
